OCR-and-Keyword-Search-with-Qwen2-VL/requirements.txt
gradio
transformers
torch
Pillow
regex
accelerate>=0.26.0
